Restructuring routes and adding multiple defendants
