+++ app/assets/sass/application.scss
@@ -650,7 +650,7 @@ legend + .form-hint {
 }
 .form-hint {
   font-family: "nta", Arial, sans-serif;
-  font-size: 16px;
+  font-size: 19px;
   line-height: 1.25;
   font-weight: 400;
   text-transform: none;
@@ -1972,4 +1972,8 @@ margin-bottom:0;
      padding-bottom:10px}
 .status-block.single{
   border:none
+}
+
+a.change {
+    margin-right: 25px;
 }
