+++ app/views/prototypes/prototype-april-2018/claimant/task-list/their-details/email/index.html
@@ -30,13 +30,15 @@ Money Claims Prototype
       {% if data.return_page and data.return_page.includes('check-and-send') or data.return_page and data.return_page.includes('check-your-answers')%}
       <form method="post" action="{{ data.return_page if data.return_page | length > 0 else '../task-list' }}" class="form">
         {% else %}
-        <form method="post" action="../../task-list" class="form">
+        <form method="post" action="defendant-add" class="form">
           {% endif %}
 
           {{formData | safe}}
 
           <h1 class="heading-large">
-            Their email address (optional)
+            {% if data['defendants'].length > 0 %}
+              <span class="form-hint">Defendant {{ data['defendants'].length + 1 }}</span>
+            {% endif %}Their email address (optional)
           </h1>
 
           <p>We'll use this to tell them you've made a claim, as well as<br> notifying them by post.</p>
@@ -53,7 +55,6 @@ Money Claims Prototype
 
 
 
-          <input type="hidden" name="their_details" value="complete">
 
         </fieldset>
